requirements.txt
mysql-connector-python==8.2.0
python-dateutil==2.8.2
